requirements.txt
aiohttp==3.9.3
aiosignal==1.3.1
annotated-types==0.6.0
anyio==4.3.0
asn1crypto==1.5.1
async-timeout==4.0.3
attrdict==2.0.1
attrs==23.2.0
base58==2.1.1
bitarray==2.9.2
blake2b-py==0.1.3
cached-property==1.5.2
cachetools==5.3.3
cargo==0.3
cbor2==5.6.2
certifi==2022.12.7
cffi==1.16.0
cfgv==3.4.0
charset-normalizer==2.0.12
ckzg==1.0.2
claripy==8.19.7.25
coincurve==19.0.1
coloredlogs==10.0
commonmark==0.9.1
configparser==6.0.1
contourpy==1.2.0
coverage==6.5.0
crytic-compile==0.3.7
cycler==0.12.1
cytoolz==0.12.3
dictionaries==0.0.2
distlib==0.3.8
distro==1.9.0
docker==7.0.0
eth-abi==2.2.0
eth-account==0.4.0
eth-bloom==1.0.4
eth-hash==0.3.3
eth-keyfile==0.5.1
eth-keys==0.2.4
eth-rlp==0.2.1
eth-typing==2.3.0
eth-utils==1.8.1
ethereum-input-decoder==0.2.2
exceptiongroup==1.2.0
filelock==3.13.1
fonttools==4.49.0
frozenlist==1.4.1
future==1.0.0
h11==0.14.0
hexbytes==0.2.3
httpcore==1.0.4
httpx==0.27.0
humanfriendly==10.0
identify==2.5.35
idna==3.4
importlib_resources==6.1.2
iniconfig==2.0.0
intervaltree==3.1.0
ipfshttpclient==0.7.0
Jinja2==3.1.3
joblib==1.3.2
jsonschema==3.2.0
jsonschema-specifications==2023.12.1
kiwisolver==1.4.5
lru-dict==1.2.0
MarkupSafe==2.0.1
matplotlib==3.8.3
mock==5.1.0
multiaddr==0.0.9
multidict==6.0.5
mypy-extensions==0.4.4
netaddr==1.3.0
nodeenv==1.8.0
numpy==1.26.4
openai==1.13.3
packaging==23.2
pakala==1.1.10
parsimonious==0.8.1
persistent==5.2
pillow==10.2.0
platformdirs==4.2.0
pluggy==1.4.0
ply==3.11
pre-commit==2.20.0
prettytable==3.10.0
protobuf==3.20.3
py-ecc==1.4.7
py-evm==0.3.0a12
py-flags==1.1.4
py-solc==3.2.0
py-solc-x==1.1.1
pycparser==2.21
pycryptodome==3.20.0
pydantic==2.6.3
pydantic_core==2.16.3
pyethash==0.1.27
pyevmasm==0.2.3
pygame==2.5.2
Pygments==2.18.0
pyparsing==2.4.7
pyrsistent==0.20.0
pysam==0.22.0
pysha3==1.0.2
PySMT==0.9.5
pytest==8.0.2
pytest-cov==4.1.0
pytest-mock==3.12.0
python-dateutil==2.9.0.post0
python-dotenv==1.0.1
pyunormalize==15.1.0
PyYAML==6.0.1
referencing==0.33.0
regex==2023.12.25
requests==2.27.1
rich==12.6.0
rlp==1.2.0
rpds-py==0.18.0
RUST==1.3.1
scikit-learn==1.4.1.post1
scipy==1.12.0
semantic-version==2.10.0
six==1.16.0
slither==0.3.0
slither-analyzer==0.10.3
sniffio==1.3.1
solc-select==1.0.4
sortedcontainers==2.4.0
threadpoolctl==3.3.0
tiktoken==0.6.0
toml==0.10.2
tomli==2.0.1
toolz==0.12.1
tqdm==4.66.2
transaction==4.0
trie==1.4.0
typing_extensions==4.10.0
urllib3==1.26.13
varint==1.0.2
virtualenv==20.25.1
wasm==1.2
wcwidth==0.2.13
web3==5.3.0
websockets==8.1
yarl==1.9.4
z3-solver==4.5.1.0.post2
zipp==3.17.0
zope.interface==6.2
